TASK-14: connection.rs -> submodules; dedupe attach-time presence logic
connection.rs (3,417 lines) becomes connection/{mod,channel_arm,
presence_arm,publish_arm}.rs. The ownership model is untouched: every state
type stays in mod.rs, owned by the loop; the arms hold only behaviour
(pub(super)/pub(crate) methods on the same structs). The conformance
ratchet now scans all four files at allowance 0.

The duplicated RTP17i re-entry block and HAS_PRESENCE attach handling
(each pasted twice in handle_attached) are now single helpers:
ConnectionCtx::reenter_internal_members and ChannelCtx::apply_has_presence.
A PendingReply enum (Publish/Op) on PendingPublish replaces the per-op
spawned oneshot-bridge tasks for presence and annotation ops; raw
91004/91005 codes replaced with ErrorCode variants.

No behavior change: suite identical before and after — unit 1246/0/26,
live integration + proxy serial 125/0/2, both ratchets green.
