+++ src/connection/mod.rs
@@ -23,13 +23,19 @@ use tokio::time::Instant;
 use crate::auth::Credential;
 use crate::error::{ErrorCode, ErrorInfo, Result};
 use crate::protocol::{
-    action, flags, ChannelEvent, ChannelMode, ChannelState, ChannelStateChange, ConnectionDetails,
-    ConnectionEvent, ConnectionState, ConnectionStateChange, ProtocolMessage,
+    action, ChannelMode, ChannelState, ChannelStateChange, ConnectionDetails, ConnectionEvent,
+    ConnectionState, ConnectionStateChange, ProtocolMessage,
 };
 use crate::rest::{AuthHeader, Format, Rest};
 use crate::transport::{Transport, TransportConnection, TransportEvent};
 
 pub(crate) type Generation = u64;
+
+mod channel_arm;
+mod presence_arm;
+mod publish_arm;
+
+use channel_arm::*;
 
 /// Everything the loop reacts to, in one totally-ordered queue.
 pub(crate) enum LoopInput {
@@ -269,7 +275,29 @@ struct PendingPublish {
     /// The wire payload, kept verbatim so an RTN19a resend reconstructs the
     /// SAME kind of ProtocolMessage (MESSAGE, PRESENCE or ANNOTATION).
     payload: PendingPayload,
-    reply: oneshot::Sender<Result<crate::rest::PublishResult>>,
+    reply: PendingReply,
+}
+
+/// The caller waiting on a pending publish's ACK/NACK. Message publishes
+/// resolve with the PublishResult; presence and annotation ops only care
+/// about success, so their result is collapsed here instead of through a
+/// per-op bridging task.
+enum PendingReply {
+    Publish(oneshot::Sender<Result<crate::rest::PublishResult>>),
+    Op(oneshot::Sender<Result<()>>),
+}
+
+impl PendingReply {
+    fn resolve(self, result: Result<crate::rest::PublishResult>) {
+        match self {
+            PendingReply::Publish(tx) => {
+                let _ = tx.send(result);
+            }
+            PendingReply::Op(tx) => {
+                let _ = tx.send(result.map(|_| ()));
+            }
+        }
+    }
 }
 
 /// The payload of a publish awaiting ACK; determines the resend pm action.
@@ -411,225 +439,6 @@ struct DeferredPresenceGet {
 struct QueuedPresenceOp {
     message: crate::rest::PresenceMessage,
     reply: oneshot::Sender<Result<()>>,
-}
-
-impl ChannelCtx {
-    /// Transition the channel state machine: snapshot first, then the event
-    /// (DESIGN.md §4 contract). RTL2g: no event when the state is unchanged.
-    fn transition(
-        &mut self,
-        to: ChannelState,
-        reason: Option<ErrorInfo>,
-        resumed: bool,
-        has_backlog: bool,
-    ) {
-        let previous = self.state;
-        if previous != to {
-            self.logger.major(|| {
-                format!(
-                    "Channel '{}': {:?} -> {:?}{}",
-                    self.name,
-                    previous,
-                    to,
-                    reason
-                        .as_ref()
-                        .map(|e| format!(" (reason: {})", e))
-                        .unwrap_or_default()
-                )
-            });
-        }
-        self.state = to;
-        if let Some(err) = &reason {
-            self.error_reason = Some(err.clone());
-        }
-        // RTL15b1: DETACHED/SUSPENDED/FAILED clear the channelSerial
-        if matches!(
-            to,
-            ChannelState::Detached | ChannelState::Suspended | ChannelState::Failed
-        ) {
-            self.channel_serial = None;
-        }
-        // RTP5a: DETACHED/FAILED clear both presence maps and fail queued
-        // presence ops + deferred gets (RTL11); RTP5f: SUSPENDED keeps the
-        // map but the sync state is no longer authoritative
-        match to {
-            ChannelState::Detached | ChannelState::Failed => {
-                self.presence.map.clear();
-                self.presence.internal.clear();
-                self.presence.sync_complete = false;
-                let err = reason.clone().unwrap_or_else(|| {
-                    ErrorInfo::new(
-                        ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                        format!("Channel became {:?}", to),
-                    )
-                });
-                for op in self.presence.queued_ops.drain(..) {
-                    let _ = op.reply.send(Err(err.clone()));
-                }
-                for get in self.presence.pending_get.drain(..) {
-                    let _ = get.reply.send(Err(err.clone()));
-                }
-            }
-            ChannelState::Suspended => {
-                self.presence.sync_complete = false;
-                let err = reason.clone().unwrap_or_else(|| {
-                    ErrorInfo::new(
-                        ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                        "Channel suspended",
-                    )
-                });
-                for op in self.presence.queued_ops.drain(..) {
-                    let _ = op.reply.send(Err(err.clone()));
-                }
-                // RTP11d: a deferred waiting get cannot complete once the
-                // presence state is out of sync
-                for get in self.presence.pending_get.drain(..) {
-                    let _ = get.reply.send(Err(ErrorInfo::with_status(
-                        91005,
-                        400,
-                        "Presence state is out of sync (channel suspended)",
-                    )));
-                }
-            }
-            _ => {}
-        }
-        self.publish_snapshot();
-        if previous != to {
-            let _ = self.events_tx.send(ChannelStateChange {
-                previous,
-                current: to,
-                event: channel_state_event(to),
-                reason,
-                resumed,
-                has_backlog,
-                retry_in: self.next_retry_in.take(),
-            });
-        }
-    }
-
-    /// RTL2g: an UPDATE event for condition changes without a state change.
-    fn emit_update(&mut self, reason: Option<ErrorInfo>, resumed: bool, has_backlog: bool) {
-        self.logger.major(|| {
-            format!(
-                "Channel '{}': UPDATE{}",
-                self.name,
-                reason
-                    .as_ref()
-                    .map(|e| format!(" (reason: {})", e))
-                    .unwrap_or_default()
-            )
-        });
-        self.publish_snapshot();
-        let _ = self.events_tx.send(ChannelStateChange {
-            previous: self.state,
-            current: self.state,
-            event: ChannelEvent::Update,
-            reason,
-            resumed,
-            has_backlog,
-            retry_in: None,
-        });
-    }
-
-    fn publish_snapshot(&self) {
-        let _ = self.snapshot_tx.send(ChannelSnapshot {
-            state: self.state,
-            options: self.options.clone(),
-            presence_sync_complete: self.presence.sync_complete,
-            error_reason: self.error_reason.clone(),
-            channel_serial: self.channel_serial.clone(),
-            attach_serial: self.attach_serial.clone(),
-            modes: self.attached_modes.clone(),
-        });
-    }
-
-    /// §8: deliver to matching subscribers; prune closed receivers.
-    fn deliver(&mut self, msg: &crate::rest::Message) {
-        self.subscribers.retain(|sub| {
-            let matches = match &sub.filter {
-                SubscriberFilter::All => true,
-                SubscriberFilter::Name(n) => msg.name.as_deref() == Some(n.as_str()),
-                SubscriberFilter::Filter(f) => f.matches(msg),
-            };
-            if !matches {
-                return true;
-            }
-            sub.sender.send(msg.clone()).is_ok()
-        });
-    }
-
-    fn resolve_attach(&mut self, result: Result<()>) {
-        for replier in self.pending_attach.drain(..) {
-            let _ = replier.send(result.clone());
-        }
-    }
-
-    fn resolve_detach(&mut self, result: Result<()>) {
-        for replier in self.pending_detach.drain(..) {
-            let _ = replier.send(result.clone());
-        }
-    }
-}
-
-fn channel_state_event(state: ChannelState) -> ChannelEvent {
-    match state {
-        ChannelState::Initialized => ChannelEvent::Initialized,
-        ChannelState::Attaching => ChannelEvent::Attaching,
-        ChannelState::Attached => ChannelEvent::Attached,
-        ChannelState::Detaching => ChannelEvent::Detaching,
-        ChannelState::Detached => ChannelEvent::Detached,
-        ChannelState::Suspended => ChannelEvent::Suspended,
-        ChannelState::Failed => ChannelEvent::Failed,
-    }
-}
-
-/// RTP6a/RTP6b: deliver to matching presence subscribers; prune closed.
-fn deliver_presence(
-    subscribers: &mut Vec<PresenceSubscriber>,
-    event: &crate::rest::PresenceMessage,
-) {
-    subscribers.retain(|sub| {
-        let matches = match &sub.actions {
-            None => true,
-            Some(actions) => event.action.map(|a| actions.contains(&a)).unwrap_or(false),
-        };
-        if !matches {
-            return true;
-        }
-        sub.sender.send(event.clone()).is_ok()
-    });
-}
-
-/// RTP11: resolve deferred gets now that the sync state is settled.
-fn resolve_presence_gets(presence: &mut PresenceCtx) {
-    for get in std::mem::take(&mut presence.pending_get) {
-        let members = presence_members(presence, &get.client_id, &get.connection_id);
-        let _ = get.reply.send(Ok(members));
-    }
-}
-
-/// RTP11c2/c3: the members list with optional clientId/connectionId filters.
-fn presence_members(
-    presence: &PresenceCtx,
-    client_id: &Option<String>,
-    connection_id: &Option<String>,
-) -> Vec<crate::rest::PresenceMessage> {
-    presence
-        .map
-        .values()
-        .into_iter()
-        .filter(|m| {
-            client_id
-                .as_ref()
-                .map(|c| m.client_id.as_deref() == Some(c.as_str()))
-                .unwrap_or(true)
-                && connection_id
-                    .as_ref()
-                    .map(|c| m.connection_id.as_deref() == Some(c.as_str()))
-                    .unwrap_or(true)
-        })
-        .cloned()
-        .collect()
 }
 
 /// All mutable connection state, owned exclusively by the loop task.
@@ -795,621 +604,6 @@ impl ConnectionCtx {
         if to == ConnectionState::Connected {
             // RTL3d/RTL4i: (re)attach channels
             self.reattach_channels_on_connected();
-        }
-    }
-
-    /// RTL6c: the publish state table. Channel SUSPENDED/FAILED and terminal
-    /// connection states fail immediately (RTL6c4); CONNECTED sends now
-    /// (RTL6c1, regardless of channel attach state, no implicit attach
-    /// RTL6c5); anything else queues per queueMessages (RTL6c2).
-    fn handle_publish(
-        &mut self,
-        name: String,
-        messages: Vec<crate::rest::Message>,
-        params: Option<serde_json::Value>,
-        reply: oneshot::Sender<Result<crate::rest::PublishResult>>,
-    ) {
-        // RTL6c4: channel state gate
-        if let Some(ch) = self.channels.get(&name) {
-            if matches!(ch.state, ChannelState::Suspended | ChannelState::Failed) {
-                let _ = reply.send(Err(ch.error_reason.clone().unwrap_or_else(|| {
-                    ErrorInfo::new(
-                        ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                        format!("Cannot publish on a {:?} channel", ch.state),
-                    )
-                })));
-                return;
-            }
-        }
-        match self.state {
-            ConnectionState::Connected => self.send_publish(name, messages, params, reply),
-            // RTL6c2: queue while a connection is plausible
-            ConnectionState::Initialized
-            | ConnectionState::Connecting
-            | ConnectionState::Disconnected => {
-                if self.rest.inner.opts.queue_messages {
-                    self.queued_publishes.push(QueuedPublish {
-                        channel: name,
-                        messages,
-                        params,
-                        reply,
-                    });
-                } else {
-                    let _ = reply.send(Err(ErrorInfo::new(
-                        ErrorCode::Disconnected.code(),
-                        "Cannot publish: not connected and queueMessages is disabled",
-                    )));
-                }
-            }
-            // RTL6c4: terminal connection states
-            _ => {
-                let _ = reply.send(Err(self.error_reason.clone().unwrap_or_else(|| {
-                    ErrorInfo::new(
-                        ErrorCode::ConnectionFailed.code(),
-                        format!("Cannot publish in connection state {:?}", self.state),
-                    )
-                })));
-            }
-        }
-    }
-
-    /// Encode for the wire (RSL4/RSL5 with the channel cipher), assign the
-    /// next msgSerial (RTN7b), send, and register the pending ACK (RTN7a).
-    fn send_publish(
-        &mut self,
-        name: String,
-        messages: Vec<crate::rest::Message>,
-        params: Option<serde_json::Value>,
-        reply: oneshot::Sender<Result<crate::rest::PublishResult>>,
-    ) {
-        let cipher = self
-            .channels
-            .get(&name)
-            .and_then(|ch| ch.options.cipher.clone());
-        let format = self.rest.inner.opts.format;
-        let mut wire_messages = Vec::with_capacity(messages.len());
-        for mut msg in messages {
-            let (data, encoding) = match crate::rest::encode_data_for_wire(
-                msg.data,
-                msg.encoding,
-                format,
-                cipher.as_ref(),
-            ) {
-                Ok(de) => de,
-                Err(e) => {
-                    let _ = reply.send(Err(e));
-                    return;
-                }
-            };
-            msg.data = data;
-            msg.encoding = encoding;
-            wire_messages.push(msg);
-        }
-        let serial = self.msg_serial;
-        self.msg_serial += 1;
-        let mut pm = ProtocolMessage::new(action::MESSAGE);
-        pm.channel = Some(name.clone());
-        pm.msg_serial = Some(serial);
-        pm.messages = Some(wire_messages.clone());
-        pm.params = params.clone();
-        self.send_protocol(pm);
-        self.pending_publishes.push(PendingPublish {
-            msg_serial: serial,
-            channel: name,
-            payload: PendingPayload::Messages {
-                messages: wire_messages,
-                params,
-            },
-            reply,
-        });
-    }
-
-    /// Resend a pending publish verbatim (RTN19a) under its (possibly
-    /// renumbered) serial.
-    fn resend_pending_publishes(&mut self) {
-        if !self.pending_publishes.is_empty() {
-            self.logger().minor(|| {
-                format!(
-                    "RTN19a: resending {} pending publish(es) on the new transport",
-                    self.pending_publishes.len()
-                )
-            });
-        }
-        let resends: Vec<ProtocolMessage> = self
-            .pending_publishes
-            .iter()
-            .map(|p| {
-                p.payload
-                    .to_protocol_message(p.channel.clone(), p.msg_serial)
-            })
-            .collect();
-        for pm in resends {
-            self.send_protocol(pm);
-        }
-    }
-
-    /// RTL6c2: queued publishes go out in order once CONNECTED.
-    fn flush_queued_publishes(&mut self) {
-        if !self.queued_publishes.is_empty() {
-            self.logger().minor(|| {
-                format!(
-                    "Flushing {} queued publish(es)",
-                    self.queued_publishes.len()
-                )
-            });
-        }
-        for q in std::mem::take(&mut self.queued_publishes) {
-            self.send_publish(q.channel, q.messages, q.params, q.reply);
-        }
-    }
-
-    /// RTN7e: fail every pending and queued publish with the given reason.
-    fn fail_all_publishes(&mut self, reason: &ErrorInfo) {
-        for p in self.pending_publishes.drain(..) {
-            let _ = p.reply.send(Err(reason.clone()));
-        }
-        for q in self.queued_publishes.drain(..) {
-            let _ = q.reply.send(Err(reason.clone()));
-        }
-    }
-
-    /// TR4s/RTL6j: an ACK resolves pending publishes with serials in
-    /// [msgSerial, msgSerial+count), pairing them with `res` entries.
-    fn handle_ack(&mut self, pm: ProtocolMessage) {
-        let first = pm.msg_serial.unwrap_or(0);
-        let count = pm.count.unwrap_or(1) as i64;
-        let res = pm.res.unwrap_or_default();
-        let acked: Vec<PendingPublish> = {
-            let mut acked = Vec::new();
-            let mut i = 0;
-            while i < self.pending_publishes.len() {
-                let serial = self.pending_publishes[i].msg_serial;
-                if serial >= first && serial < first + count {
-                    acked.push(self.pending_publishes.remove(i));
-                } else {
-                    i += 1;
-                }
-            }
-            acked
-        };
-        if acked.is_empty() {
-            self.logger().error(|| {
-                format!(
-                    "ACK for unknown msgSerial range [{}, {}) — no pending operation matches",
-                    first,
-                    first + count
-                )
-            });
-        }
-        for p in acked {
-            let idx = (p.msg_serial - first) as usize;
-            let result = res
-                .get(idx)
-                .map(|r| crate::rest::PublishResult {
-                    serials: r.serials.clone(),
-                    message_id: None,
-                })
-                .unwrap_or_default();
-            let _ = p.reply.send(Ok(result));
-        }
-    }
-
-    /// A NACK fails the addressed pending publishes (RTL6j).
-    fn handle_nack(&mut self, pm: ProtocolMessage) {
-        let first = pm.msg_serial.unwrap_or(0);
-        let count = pm.count.unwrap_or(1) as i64;
-        let reason = pm.error.unwrap_or_else(|| {
-            ErrorInfo::with_status(ErrorCode::InternalError.code(), 500, "Publish rejected")
-        });
-        let mut i = 0;
-        let mut matched = false;
-        while i < self.pending_publishes.len() {
-            let serial = self.pending_publishes[i].msg_serial;
-            if serial >= first && serial < first + count {
-                let p = self.pending_publishes.remove(i);
-                self.logger()
-                    .minor(|| format!("NACK for msgSerial {}: {}", serial, reason));
-                let _ = p.reply.send(Err(reason.clone()));
-                matched = true;
-            } else {
-                i += 1;
-            }
-        }
-        if !matched {
-            self.logger().error(|| {
-                format!(
-                    "NACK for unknown msgSerial range [{}, {}) — no pending operation matches",
-                    first,
-                    first + count
-                )
-            });
-        }
-    }
-
-    /// RTP8/9/10: a presence operation per the RTP16 connection/channel
-    /// state table: send when ATTACHED, queue while ATTACHING (or implicit
-    /// attach from INITIALIZED, RTP8d), error otherwise (RTP8g/RTP16c).
-    fn handle_presence_op(
-        &mut self,
-        name: String,
-        message: crate::rest::PresenceMessage,
-        reply: oneshot::Sender<Result<()>>,
-    ) {
-        let connected = self.state == ConnectionState::Connected;
-        let _rtt = self.rest.inner.opts.realtime_request_timeout;
-        let Some(ch) = self.channels.get_mut(&name) else {
-            let _ = reply.send(Err(ErrorInfo::new(
-                ErrorCode::BadRequest.code(),
-                "Unknown channel",
-            )));
-            return;
-        };
-        match ch.state {
-            ChannelState::Attached => {
-                if connected {
-                    self.send_presence(name, message, reply);
-                } else if self.rest.inner.opts.queue_messages
-                    && matches!(
-                        self.state,
-                        ConnectionState::Connecting | ConnectionState::Disconnected
-                    )
-                {
-                    ch.presence
-                        .queued_ops
-                        .push(QueuedPresenceOp { message, reply });
-                } else {
-                    let _ = reply.send(Err(ErrorInfo::new(
-                        ErrorCode::Disconnected.code(),
-                        format!("Cannot send presence in connection state {:?}", self.state),
-                    )));
-                }
-            }
-            // RTP16b: queued until the attach completes
-            ChannelState::Attaching => {
-                ch.presence
-                    .queued_ops
-                    .push(QueuedPresenceOp { message, reply });
-            }
-            // RTP8d: an INITIALIZED channel is implicitly attached
-            ChannelState::Initialized => {
-                ch.presence
-                    .queued_ops
-                    .push(QueuedPresenceOp { message, reply });
-                let (attach_reply, _rx) = oneshot::channel();
-                self.handle_attach(name, attach_reply);
-            }
-            // RTP8g/RTP16c: DETACHED/DETACHING/SUSPENDED/FAILED error
-            _ => {
-                let _ = reply.send(Err(ErrorInfo::with_status(
-                    ErrorCode::UnableToEnterPresenceChannelInvalidChannelState.code(),
-                    400,
-                    format!("Cannot send presence in channel state {:?}", ch.state),
-                )));
-            }
-        }
-    }
-
-    /// Send a PRESENCE ProtocolMessage with the next msgSerial; the ACK/NACK
-    /// resolves the reply through the pending-publish machinery (RTL11a:
-    /// resolution is unaffected by later channel state changes).
-    fn send_presence(
-        &mut self,
-        name: String,
-        message: crate::rest::PresenceMessage,
-        reply: oneshot::Sender<Result<()>>,
-    ) {
-        let serial = self.msg_serial;
-        self.msg_serial += 1;
-        let mut pm = ProtocolMessage::new(action::PRESENCE);
-        pm.channel = Some(name.clone());
-        pm.msg_serial = Some(serial);
-        pm.presence = Some(vec![message.clone()]);
-        self.send_protocol(pm);
-        let (ack_reply, ack_rx) = oneshot::channel::<Result<crate::rest::PublishResult>>();
-        self.pending_publishes.push(PendingPublish {
-            msg_serial: serial,
-            channel: name,
-            payload: PendingPayload::Presence(vec![message]),
-            reply: ack_reply,
-        });
-        tokio::spawn(async move {
-            let outcome = match ack_rx.await {
-                Ok(Ok(_)) => Ok(()),
-                Ok(Err(e)) => Err(e),
-                Err(_) => Err(ErrorInfo::new(
-                    ErrorCode::Disconnected.code(),
-                    "Connection loop dropped the presence op",
-                )),
-            };
-            let _ = reply.send(outcome);
-        });
-    }
-
-    /// RTP11: presence get with waitForSync semantics.
-    fn handle_presence_get(
-        &mut self,
-        name: String,
-        wait_for_sync: bool,
-        client_id: Option<String>,
-        connection_id: Option<String>,
-        reply: oneshot::Sender<Result<Vec<crate::rest::PresenceMessage>>>,
-    ) {
-        let Some(ch) = self.channels.get_mut(&name) else {
-            let _ = reply.send(Ok(Vec::new()));
-            return;
-        };
-        // RTP11d: SUSPENDED errors unless waitForSync=false
-        if ch.state == ChannelState::Suspended {
-            if wait_for_sync {
-                let _ = reply.send(Err(ErrorInfo::with_status(
-                    91005,
-                    400,
-                    "Presence state is out of sync (channel suspended)",
-                )));
-            } else {
-                let _ = reply.send(Ok(presence_members(
-                    &ch.presence,
-                    &client_id,
-                    &connection_id,
-                )));
-            }
-            return;
-        }
-        if !wait_for_sync || ch.presence.sync_complete {
-            let _ = reply.send(Ok(presence_members(
-                &ch.presence,
-                &client_id,
-                &connection_id,
-            )));
-            return;
-        }
-        // RTP11a/RTP11b: defer until the sync completes (the implicit attach
-        // is issued handle-side)
-        ch.presence.pending_get.push(DeferredPresenceGet {
-            client_id,
-            connection_id,
-            reply,
-        });
-    }
-
-    /// RTAN1b: annotation ops share the message-publish state table; the
-    /// wire shape is an ANNOTATION ProtocolMessage resolved via ACK/NACK
-    /// (RTAN1d).
-    fn handle_annotation_op(
-        &mut self,
-        name: String,
-        annotation: crate::rest::Annotation,
-        reply: oneshot::Sender<Result<()>>,
-    ) {
-        // RTL6c4-shaped channel gate
-        if let Some(ch) = self.channels.get(&name) {
-            if matches!(ch.state, ChannelState::Suspended | ChannelState::Failed) {
-                let _ = reply.send(Err(ch.error_reason.clone().unwrap_or_else(|| {
-                    ErrorInfo::new(
-                        ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                        format!("Cannot publish an annotation on a {:?} channel", ch.state),
-                    )
-                })));
-                return;
-            }
-        }
-        if self.state != ConnectionState::Connected {
-            let _ = reply.send(Err(ErrorInfo::new(
-                ErrorCode::Disconnected.code(),
-                format!(
-                    "Cannot publish an annotation in connection state {:?}",
-                    self.state
-                ),
-            )));
-            return;
-        }
-        // RTAN1a/RSAN1c3: annotation data is encoded per RSL4 (annotations
-        // are not encrypted, so no cipher applies)
-        let mut annotation = annotation;
-        let format = self.rest.inner.opts.format;
-        match crate::rest::encode_data_for_wire(
-            annotation.data.clone(),
-            annotation.encoding.clone(),
-            format,
-            None,
-        ) {
-            Ok((data, encoding)) => {
-                annotation.data = data;
-                annotation.encoding = encoding;
-            }
-            Err(e) => {
-                let _ = reply.send(Err(e));
-                return;
-            }
-        }
-        let serial = self.msg_serial;
-        self.msg_serial += 1;
-        let mut pm = ProtocolMessage::new(action::ANNOTATION);
-        pm.channel = Some(name.clone());
-        pm.msg_serial = Some(serial);
-        pm.annotations = Some(vec![annotation.clone()]);
-        self.send_protocol(pm);
-        let (ack_reply, ack_rx) = oneshot::channel::<Result<crate::rest::PublishResult>>();
-        self.pending_publishes.push(PendingPublish {
-            msg_serial: serial,
-            channel: name,
-            payload: PendingPayload::Annotations(vec![annotation]),
-            reply: ack_reply,
-        });
-        tokio::spawn(async move {
-            let outcome = match ack_rx.await {
-                Ok(Ok(_)) => Ok(()),
-                Ok(Err(e)) => Err(e),
-                Err(_) => Err(ErrorInfo::new(
-                    ErrorCode::Disconnected.code(),
-                    "Connection loop dropped the annotation op",
-                )),
-            };
-            let _ = reply.send(outcome);
-        });
-    }
-
-    /// RTAN4: inbound ANNOTATION — decode entries and dispatch to matching
-    /// subscribers (RTAN4c type filters).
-    fn handle_annotation_action(&mut self, pm: ProtocolMessage) {
-        self.update_channel_serial(&pm);
-        let Some(name) = pm.channel.clone() else {
-            return;
-        };
-        let Some(ch) = self.channels.get_mut(&name) else {
-            return;
-        };
-        if ch.state != ChannelState::Attached {
-            return;
-        }
-        let entries = pm.annotations.clone().unwrap_or_default();
-        for (index, mut ann) in entries.into_iter().enumerate() {
-            if ann.id.is_none() {
-                if let Some(pm_id) = &pm.id {
-                    ann.id = Some(format!("{}:{}", pm_id, index));
-                }
-            }
-            if ann.timestamp.is_none() {
-                ann.timestamp = pm.timestamp;
-            }
-            // RTAN4b1: annotation data decodes per RSL6 (no cipher —
-            // annotations are not encrypted)
-            let (data, encoding) = crate::rest::decode_data(ann.data, ann.encoding, None);
-            ann.data = data;
-            ann.encoding = encoding;
-            ch.annotation_subscribers.retain(|sub| {
-                let matches = sub
-                    .type_filter
-                    .as_ref()
-                    .map(|t| ann.annotation_type.as_deref() == Some(t.as_str()))
-                    .unwrap_or(true);
-                if !matches {
-                    return true;
-                }
-                sub.sender.send(ann.clone()).is_ok()
-            });
-        }
-    }
-
-    /// RTL15b: MESSAGE/PRESENCE/ANNOTATION carrying a channelSerial update the
-    /// channel's serial (SYNC is excluded — see handle_presence_action).
-    fn update_channel_serial(&mut self, pm: &ProtocolMessage) {
-        let Some(name) = &pm.channel else { return };
-        let Some(serial) = &pm.channel_serial else {
-            return;
-        };
-        if let Some(ch) = self.channels.get_mut(name) {
-            ch.channel_serial = Some(serial.clone());
-            ch.publish_snapshot();
-        }
-    }
-
-    /// A MESSAGE from the server: TM2 field population, RSL6 decode with the
-    /// channel cipher, RTL17 attached-only delivery, subscriber dispatch (§8).
-    fn handle_message_action(&mut self, pm: ProtocolMessage) {
-        self.update_channel_serial(&pm);
-        let Some(name) = pm.channel.clone() else {
-            return;
-        };
-        let Some(ch) = self.channels.get_mut(&name) else {
-            return;
-        };
-        // RTL17: messages are only delivered while ATTACHED
-        if ch.state != ChannelState::Attached {
-            ch.logger.minor(|| {
-                format!(
-                    "RTL17: dropping MESSAGE for channel '{}' in state {:?}",
-                    name, ch.state
-                )
-            });
-            return;
-        }
-        let wire = pm.messages.clone().unwrap_or_default();
-        for (index, mut msg) in wire.into_iter().enumerate() {
-            // TM2a: id defaults to protocolMessage.id + ":" + index
-            if msg.id.is_none() {
-                if let Some(pm_id) = &pm.id {
-                    msg.id = Some(format!("{}:{}", pm_id, index));
-                }
-            }
-            // TM2c: connectionId inherited unless already present
-            if msg.connection_id.is_none() {
-                msg.connection_id = pm.connection_id.clone();
-            }
-            // TM2f: timestamp inherited unless already present
-            if msg.timestamp.is_none() {
-                msg.timestamp = pm.timestamp;
-            }
-            // RSL6: decode/decrypt with the channel cipher (also applies
-            // the TM2s version defaulting, after the inheritance above)
-            msg.decode_with_cipher(ch.options.cipher.as_ref());
-            ch.deliver(&msg);
-        }
-    }
-
-    /// RTP6/RTP17/RTP18/RTP19: inbound PRESENCE or SYNC. Field population
-    /// follows TM2 conventions; events are dispatched per RTP2 newness.
-    fn handle_presence_action(&mut self, pm: ProtocolMessage, is_sync: bool) {
-        // RTL15b: PRESENCE updates the channel serial; SYNC does not — its
-        // channelSerial carries the sync cursor ("<sequence>:<cursor>"), which
-        // is not a channel serial and would be rejected by the server if sent
-        // back in a reattach ATTACH (RTL4c1).
-        if !is_sync {
-            self.update_channel_serial(&pm);
-        }
-        let Some(name) = pm.channel.clone() else {
-            return;
-        };
-        let own_connection = self.id.clone();
-        let Some(ch) = self.channels.get_mut(&name) else {
-            return;
-        };
-
-        if is_sync && !ch.presence.map.sync_in_progress() {
-            // RTP18a: a new sync page stream begins
-            ch.logger
-                .minor(|| format!("Channel '{}': presence SYNC started", name));
-            ch.presence.map.start_sync();
-            ch.presence.sync_complete = false;
-            ch.publish_snapshot();
-        }
-
-        let wire = pm.presence.clone().unwrap_or_default();
-        for (index, mut msg) in wire.into_iter().enumerate() {
-            // TM2-shaped inheritance
-            if msg.id.is_none() {
-                if let Some(pm_id) = &pm.id {
-                    msg.id = Some(format!("{}:{}", pm_id, index));
-                }
-            }
-            if msg.connection_id.is_none() {
-                msg.connection_id = pm.connection_id.clone();
-            }
-            if msg.timestamp.is_none() {
-                msg.timestamp = pm.timestamp;
-            }
-            msg.decode_with_cipher(ch.options.cipher.as_ref());
-            // RTP17: members entered through THIS connection feed the
-            // internal map
-            if own_connection.is_some() && msg.connection_id == own_connection {
-                ch.presence.internal.put(&msg);
-            }
-            if let Some(event) = ch.presence.map.put(&msg) {
-                deliver_presence(&mut ch.presence.subscribers, &event);
-            }
-        }
-
-        // RTP18b/RTP18c: the sync completes when the cursor is exhausted
-        if is_sync && !crate::presence::sync_continues(&pm.channel_serial) {
-            ch.logger
-                .minor(|| format!("Channel '{}': presence SYNC complete", name));
-            let leaves = ch.presence.map.end_sync();
-            for leave in &leaves {
-                deliver_presence(&mut ch.presence.subscribers, leave);
-            }
-            ch.presence.sync_complete = true;
-            ch.publish_snapshot();
-            resolve_presence_gets(&mut ch.presence);
         }
     }
 
@@ -1879,8 +1073,11 @@ impl ConnectionCtx {
                 });
                 if let Some(ch) = self.channels.get_mut(&name) {
                     // RTP17e: 91004 wraps the underlying failure
-                    let mut wrapped =
-                        ErrorInfo::with_cause(91004, "Automatic presence re-entry failed", error);
+                    let mut wrapped = ErrorInfo::with_cause(
+                        ErrorCode::UnableToAutomaticallyReEnterPresenceChannel.code(),
+                        "Automatic presence re-entry failed",
+                        error,
+                    );
                     wrapped.status_code = Some(400);
                     // RTP17e: resumed=true — the channel itself was continuous
                     ch.emit_update(Some(wrapped), true, false);
@@ -2423,484 +1620,6 @@ impl ConnectionCtx {
 
     // --- Channel lifecycle (RTL2/RTL3/RTL4/RTL5, DESIGN.md §7) ---
 
-    /// RTL4: attach a channel.
-    fn handle_attach(&mut self, name: String, reply: oneshot::Sender<Result<()>>) {
-        let conn_state = self.state;
-        let rtt = self.rest.inner.opts.realtime_request_timeout;
-        let Some(ch) = self.channels.get_mut(&name) else {
-            let _ = reply.send(Err(ErrorInfo::new(
-                ErrorCode::ChannelOperationFailed.code(),
-                "Channel has been released",
-            )));
-            return;
-        };
-        match ch.state {
-            // RTL4a: already attached — immediate success
-            ChannelState::Attached => {
-                let _ = reply.send(Ok(()));
-            }
-            // RTL4h: attach in progress — share its outcome
-            ChannelState::Attaching => {
-                ch.pending_attach.push(reply);
-            }
-            // RTL4h: detaching — attach once the detach completes
-            ChannelState::Detaching => {
-                ch.attach_pending = true;
-                ch.pending_attach.push(reply);
-            }
-            // RTL4g covers Failed (proceeds, clearing errorReason via RTL4c)
-            ChannelState::Initialized
-            | ChannelState::Detached
-            | ChannelState::Suspended
-            | ChannelState::Failed => match conn_state {
-                // RTL4b: invalid connection states
-                ConnectionState::Closing
-                | ConnectionState::Closed
-                | ConnectionState::Failed
-                | ConnectionState::Suspended => {
-                    let _ = reply.send(Err(ErrorInfo::new(
-                        ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                        format!("Cannot attach while the connection is {:?}", conn_state),
-                    )));
-                }
-                // RTL4i: queue until the connection is CONNECTED
-                ConnectionState::Initialized
-                | ConnectionState::Connecting
-                | ConnectionState::Disconnected => {
-                    // RTL4c: a new attach clears errorReason
-                    ch.error_reason = None;
-                    ch.pending_attach.push(reply);
-                    ch.attach_pending = true;
-                    ch.transition(ChannelState::Attaching, None, false, false);
-                }
-                ConnectionState::Connected => {
-                    ch.error_reason = None;
-                    ch.pending_attach.push(reply);
-                    ch.transition(ChannelState::Attaching, None, false, false);
-                    ch.op_deadline = Some(Instant::now() + rtt);
-                    let msg = attach_message(ch);
-                    self.send_protocol(msg);
-                }
-            },
-        }
-    }
-
-    /// RTL5: detach a channel.
-    fn handle_detach(&mut self, name: String, reply: oneshot::Sender<Result<()>>) {
-        let conn_state = self.state;
-        let rtt = self.rest.inner.opts.realtime_request_timeout;
-        let Some(ch) = self.channels.get_mut(&name) else {
-            let _ = reply.send(Ok(()));
-            return;
-        };
-        match ch.state {
-            // RTL5a: nothing to detach
-            ChannelState::Initialized | ChannelState::Detached => {
-                let _ = reply.send(Ok(()));
-            }
-            // RTL5b: detach from FAILED is an error
-            ChannelState::Failed => {
-                let _ = reply.send(Err(ErrorInfo::new(
-                    ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                    "Cannot detach a failed channel",
-                )));
-            }
-            // RTL5j: suspended → detached immediately
-            ChannelState::Suspended => {
-                let _ = reply.send(Ok(()));
-                ch.transition(ChannelState::Detached, None, false, false);
-            }
-            // RTL5i: detach in progress — share its outcome
-            ChannelState::Detaching => {
-                ch.pending_detach.push(reply);
-            }
-            // RTL5i: attaching — detach once the attach completes
-            ChannelState::Attaching => {
-                if conn_state == ConnectionState::Connected {
-                    ch.detach_pending = true;
-                    ch.pending_detach.push(reply);
-                } else {
-                    // RTL5l: no live connection — abandon the queued attach
-                    // and go straight to DETACHED, nothing on the wire
-                    ch.attach_pending = false;
-                    ch.op_deadline = None;
-                    ch.resolve_attach(Err(ErrorInfo::new(
-                        ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                        "Attach superseded by detach",
-                    )));
-                    let _ = reply.send(Ok(()));
-                    ch.transition(ChannelState::Detached, None, false, false);
-                }
-            }
-            ChannelState::Attached => {
-                if conn_state == ConnectionState::Connected {
-                    // RTL5d: DETACH on the wire, await DETACHED
-                    ch.op_revert_state = ch.state;
-                    ch.pending_detach.push(reply);
-                    ch.transition(ChannelState::Detaching, None, false, false);
-                    ch.op_deadline = Some(Instant::now() + rtt);
-                    let mut msg = ProtocolMessage::new(action::DETACH);
-                    msg.channel = Some(ch.name.clone());
-                    self.send_protocol(msg);
-                } else {
-                    // RTL5l: no live connection — detached immediately
-                    let _ = reply.send(Ok(()));
-                    ch.transition(ChannelState::Detached, None, false, false);
-                }
-            }
-        }
-    }
-
-    /// ATTACHED received from the server.
-    fn handle_attached(&mut self, pm: ProtocolMessage) {
-        let Some(name) = pm.channel.clone() else {
-            return;
-        };
-        let rtt = self.rest.inner.opts.realtime_request_timeout;
-        let Some(ch) = self.channels.get_mut(&name) else {
-            return;
-        };
-        let resumed = pm.flags.map(|f| f & flags::RESUMED != 0).unwrap_or(false);
-        let has_backlog = pm
-            .flags
-            .map(|f| f & flags::HAS_BACKLOG != 0)
-            .unwrap_or(false);
-        match ch.state {
-            ChannelState::Attaching => {
-                ch.attach_serial = pm.channel_serial.clone();
-                ch.channel_serial = pm.channel_serial.clone();
-                // RTL4m: modes granted by the server
-                ch.attached_modes = pm.flags.map(modes_from_flags);
-                ch.has_been_attached = true;
-                ch.op_deadline = None;
-                // RTL13b: a successful attach ends the retry cycle
-                ch.retry_at = None;
-                ch.retry_count = 0;
-                // RTP1/RTP19a: HAS_PRESENCE announces an incoming sync;
-                // without it the presence set is authoritatively empty
-                let has_presence = pm
-                    .flags
-                    .map(|f| f & flags::HAS_PRESENCE != 0)
-                    .unwrap_or(false);
-                if has_presence {
-                    ch.presence.map.start_sync();
-                    ch.presence.sync_complete = false;
-                } else {
-                    ch.presence.map.start_sync();
-                    let leaves = ch.presence.map.end_sync();
-                    for leave in &leaves {
-                        deliver_presence(&mut ch.presence.subscribers, leave);
-                    }
-                    ch.presence.sync_complete = true;
-                    resolve_presence_gets(&mut ch.presence);
-                }
-                ch.resolve_attach(Ok(()));
-                ch.transition(ChannelState::Attached, pm.error, resumed, has_backlog);
-                // RTP5b: queued presence ops go out now
-                let queued: Vec<QueuedPresenceOp> = ch.presence.queued_ops.drain(..).collect();
-                // RTP17i: automatic re-entry of internal members on a
-                // non-resumed attach; RTP17g1: the id is omitted when the
-                // connectionId changed
-                let mut reentries = Vec::new();
-                if !resumed {
-                    let own = self.id.clone();
-                    for member in ch.presence.internal.values() {
-                        let mut enter = member.clone();
-                        enter.action = Some(crate::rest::PresenceAction::Enter);
-                        if enter.connection_id != own {
-                            enter.id = None;
-                        }
-                        enter.connection_id = None;
-                        reentries.push(enter);
-                    }
-                }
-                let detach_now = std::mem::take(&mut ch.detach_pending);
-                for op in queued {
-                    self.send_presence(name.clone(), op.message, op.reply);
-                }
-                for enter in reentries {
-                    let (reply, rx) = oneshot::channel();
-                    self.send_presence(name.clone(), enter, reply);
-                    // RTP17e: a failed re-entry surfaces as a channel UPDATE
-                    // with the error
-                    let input_tx = self.input_tx.clone();
-                    let chan = name.clone();
-                    tokio::spawn(async move {
-                        if let Ok(Err(err)) = rx.await {
-                            let _ = input_tx.send(LoopInput::Cmd(Command::PresenceReentryFailed {
-                                name: chan,
-                                error: err,
-                            }));
-                        }
-                    });
-                }
-                // RTL5i: a queued detach proceeds now
-                if detach_now {
-                    let (tx, _rx) = oneshot::channel();
-                    self.handle_detach(name, tx);
-                }
-            }
-            ChannelState::Attached => {
-                // RTL12-shaped: an additional ATTACHED is an UPDATE
-                ch.attach_serial = pm.channel_serial.clone();
-                ch.channel_serial = pm.channel_serial.clone();
-                if let Some(f) = pm.flags {
-                    ch.attached_modes = Some(modes_from_flags(f));
-                }
-                // RTL12: RESUMED means continuity was preserved — no UPDATE
-                if !resumed {
-                    ch.emit_update(pm.error, resumed, has_backlog);
-                    // RTP1/RTP19a: the flagless re-ATTACHED makes the
-                    // presence set authoritatively empty; with HAS_PRESENCE a
-                    // fresh sync follows
-                    let has_presence = pm
-                        .flags
-                        .map(|f| f & flags::HAS_PRESENCE != 0)
-                        .unwrap_or(false);
-                    if has_presence {
-                        ch.presence.map.start_sync();
-                        ch.presence.sync_complete = false;
-                        ch.publish_snapshot();
-                    } else {
-                        ch.presence.map.start_sync();
-                        let leaves = ch.presence.map.end_sync();
-                        for leave in &leaves {
-                            deliver_presence(&mut ch.presence.subscribers, leave);
-                        }
-                        ch.presence.sync_complete = true;
-                        ch.publish_snapshot();
-                        resolve_presence_gets(&mut ch.presence);
-                    }
-                    // RTP17i: continuity was lost — re-enter internal members
-                    let own = self.id.clone();
-                    let mut reentries = Vec::new();
-                    for member in ch.presence.internal.values() {
-                        let mut enter = member.clone();
-                        enter.action = Some(crate::rest::PresenceAction::Enter);
-                        if enter.connection_id != own {
-                            enter.id = None; // RTP17g1
-                        }
-                        enter.connection_id = None;
-                        reentries.push(enter);
-                    }
-                    for enter in reentries {
-                        let (reply, rx) = oneshot::channel();
-                        self.send_presence(name.clone(), enter, reply);
-                        let input_tx = self.input_tx.clone();
-                        let chan = name.clone();
-                        tokio::spawn(async move {
-                            if let Ok(Err(err)) = rx.await {
-                                let _ =
-                                    input_tx.send(LoopInput::Cmd(Command::PresenceReentryFailed {
-                                        name: chan,
-                                        error: err,
-                                    }));
-                            }
-                        });
-                    }
-                }
-            }
-            // RTL5k: an ATTACHED while detaching/detached is answered with DETACH
-            ChannelState::Detaching | ChannelState::Detached => {
-                ch.op_deadline = Some(Instant::now() + rtt);
-                let mut msg = ProtocolMessage::new(action::DETACH);
-                msg.channel = Some(name);
-                self.send_protocol(msg);
-            }
-            _ => {}
-        }
-    }
-
-    /// DETACHED received from the server.
-    fn handle_detached(&mut self, pm: ProtocolMessage) {
-        let Some(name) = pm.channel.clone() else {
-            return;
-        };
-        let Some(ch) = self.channels.get_mut(&name) else {
-            return;
-        };
-        match ch.state {
-            ChannelState::Detaching => {
-                ch.op_deadline = None;
-                ch.resolve_detach(Ok(()));
-                ch.transition(ChannelState::Detached, pm.error, false, false);
-                if ch.release_on_detach {
-                    if let Some(reply) = ch.release_reply.take() {
-                        let _ = reply.send(());
-                    }
-                    self.channels.remove(&name);
-                    return;
-                }
-                // RTL4h: a queued attach proceeds now
-                let attach_now =
-                    std::mem::take(&mut self.channels.get_mut(&name).unwrap().attach_pending);
-                if attach_now {
-                    let (tx, _rx) = oneshot::channel();
-                    self.handle_attach(name, tx);
-                }
-            }
-            // RTL13a: server-initiated DETACHED on an ATTACHED or SUSPENDED
-            // channel triggers an immediate reattach
-            ChannelState::Attached | ChannelState::Suspended => {
-                let rtt = self.rest.inner.opts.realtime_request_timeout;
-                let Some(ch) = self.channels.get_mut(&name) else {
-                    return;
-                };
-                ch.transition(ChannelState::Attaching, pm.error, false, false);
-                ch.op_deadline = Some(Instant::now() + rtt);
-                let msg = attach_message(ch);
-                self.send_protocol(msg);
-            }
-            // RTL13b: DETACHED while ATTACHING is a failed (re)attach — go
-            // SUSPENDED and schedule a retry
-            ChannelState::Attaching => {
-                let reason = pm.error.clone();
-                if let Some(ch) = self.channels.get_mut(&name) {
-                    ch.op_deadline = None;
-                    ch.resolve_attach(Err(reason.clone().unwrap_or_else(|| {
-                        ErrorInfo::new(
-                            ErrorCode::ChannelOperationFailedInvalidChannelState.code(),
-                            "Attach rejected by the server",
-                        )
-                    })));
-                }
-                self.suspend_channel_with_retry(&name, reason);
-            }
-            _ => {}
-        }
-    }
-
-    /// ERROR with a channel set: the attach/detach failed (RTL4e/RTL5e-shaped).
-    fn handle_channel_error(&mut self, pm: ProtocolMessage) {
-        let Some(name) = pm.channel.clone() else {
-            return;
-        };
-        let Some(ch) = self.channels.get_mut(&name) else {
-            return;
-        };
-        ch.op_deadline = None;
-        let err = pm.error.clone().unwrap_or_else(|| {
-            ErrorInfo::new(ErrorCode::ChannelOperationFailed.code(), "Channel error")
-        });
-        ch.resolve_attach(Err(err.clone()));
-        ch.resolve_detach(Err(err.clone()));
-        ch.transition(ChannelState::Failed, Some(err), false, false);
-    }
-
-    /// RTL3: connection-state side effects on channels — applied atomically
-    /// with the connection transition (DESIGN.md §7).
-    fn apply_connection_effects_to_channels(
-        &mut self,
-        conn_state: ConnectionState,
-        reason: &Option<ErrorInfo>,
-    ) {
-        match conn_state {
-            // RTL3a: FAILED fails attached/attaching channels
-            ConnectionState::Failed => {
-                for ch in self.channels.values_mut() {
-                    if matches!(ch.state, ChannelState::Attached | ChannelState::Attaching) {
-                        ch.op_deadline = None;
-                        ch.resolve_attach(Err(reason.clone().unwrap_or_else(|| {
-                            ErrorInfo::new(ErrorCode::ConnectionFailed.code(), "Connection failed")
-                        })));
-                        ch.transition(ChannelState::Failed, reason.clone(), false, false);
-                    }
-                }
-            }
-            // RTL3b: CLOSED detaches attached/attaching channels
-            ConnectionState::Closed => {
-                for ch in self.channels.values_mut() {
-                    if matches!(ch.state, ChannelState::Attached | ChannelState::Attaching) {
-                        ch.op_deadline = None;
-                        ch.resolve_attach(Err(ErrorInfo::new(
-                            ErrorCode::ConnectionClosed.code(),
-                            "Connection closed",
-                        )));
-                        ch.transition(ChannelState::Detached, None, false, false);
-                    }
-                }
-            }
-            // RTL3c: SUSPENDED suspends attached/attaching channels
-            ConnectionState::Suspended => {
-                for ch in self.channels.values_mut() {
-                    if matches!(ch.state, ChannelState::Attached | ChannelState::Attaching) {
-                        ch.op_deadline = None;
-                        ch.resolve_attach(Err(reason.clone().unwrap_or_else(|| {
-                            ErrorInfo::new(
-                                ErrorCode::ConnectionSuspended.code(),
-                                "Connection suspended",
-                            )
-                        })));
-                        ch.transition(ChannelState::Suspended, reason.clone(), false, false);
-                    }
-                }
-            }
-            // RTL3e: DISCONNECTED leaves channel states untouched
-            _ => {}
-        }
-        // RTL13c: channel reattach retries only run while CONNECTED
-        if conn_state != ConnectionState::Connected {
-            for ch in self.channels.values_mut() {
-                ch.retry_at = None;
-            }
-        }
-    }
-
-    /// RTL13b: transition a channel to SUSPENDED and schedule the next
-    /// reattach retry (RTB1 backoff over channelRetryTimeout), provided the
-    /// connection is still CONNECTED (RTL13c).
-    fn suspend_channel_with_retry(&mut self, name: &str, reason: Option<ErrorInfo>) {
-        let connected = self.state == ConnectionState::Connected;
-        let base = self.rest.inner.opts.channel_retry_timeout;
-        let Some(ch) = self.channels.get_mut(name) else {
-            return;
-        };
-        if connected {
-            let delay = retry_delay(base, ch.retry_count);
-            ch.retry_count += 1;
-            ch.retry_at = Some(Instant::now() + delay);
-            ch.next_retry_in = Some(delay);
-            ch.logger.minor(|| {
-                format!(
-                    "Channel '{}': scheduling reattach retry {} in {:?} (RTL13b)",
-                    name, ch.retry_count, delay
-                )
-            });
-        }
-        ch.transition(ChannelState::Suspended, reason, false, false);
-    }
-
-    /// RTL3d: on CONNECTED, (re)attach channels that were attached, attaching,
-    /// suspended, or queued (RTL4i).
-    fn reattach_channels_on_connected(&mut self) {
-        let rtt = self.rest.inner.opts.realtime_request_timeout;
-        let mut to_send = Vec::new();
-        for ch in self.channels.values_mut() {
-            let queued = std::mem::take(&mut ch.attach_pending);
-            let needs_attach = queued
-                || matches!(
-                    ch.state,
-                    ChannelState::Attached | ChannelState::Attaching | ChannelState::Suspended
-                );
-            if needs_attach {
-                if ch.state != ChannelState::Attaching {
-                    ch.transition(ChannelState::Attaching, None, false, false);
-                }
-                ch.op_deadline = Some(Instant::now() + rtt);
-                to_send.push(attach_message(ch));
-            } else if ch.state == ChannelState::Detaching {
-                // RTN19b: a pending DETACH is resent on the new transport
-                ch.op_deadline = Some(Instant::now() + rtt);
-                let mut msg = ProtocolMessage::new(action::DETACH);
-                msg.channel = Some(ch.name.clone());
-                to_send.push(msg);
-            }
-        }
-        for msg in to_send {
-            self.send_protocol(msg);
-        }
-    }
-
     fn can_renew_token(&self) -> bool {
         let cfg = self.rest.auth_config();
         cfg.callback.is_some() || cfg.url.is_some() || cfg.key.is_some()
@@ -3083,71 +1802,6 @@ impl ConnectionCtx {
             }
         }
     }
-}
-
-/// RTL4c/RTL4c1/RTL4k/RTL4l/RTL4j: build the ATTACH message for a channel.
-fn attach_message(ch: &ChannelCtx) -> ProtocolMessage {
-    let mut msg = ProtocolMessage::new(action::ATTACH);
-    msg.channel = Some(ch.name.clone());
-    // RTL4c1: include the channelSerial from the previous attachment
-    if let Some(serial) = &ch.channel_serial {
-        msg.channel_serial = Some(serial.clone());
-    }
-    // RTL4k: requested channel params
-    if !ch.options.params.is_empty() {
-        let map: serde_json::Map<String, serde_json::Value> = ch
-            .options
-            .params
-            .iter()
-            .map(|(k, v)| (k.clone(), serde_json::Value::String(v.clone())))
-            .collect();
-        msg.params = Some(serde_json::Value::Object(map));
-    }
-    // RTL4l: requested modes as flags; RTL4j: ATTACH_RESUME on reattach
-    let mut flag_bits: u64 = ch
-        .options
-        .modes
-        .iter()
-        .map(|m| match m {
-            ChannelMode::Presence => flags::PRESENCE,
-            ChannelMode::Publish => flags::PUBLISH,
-            ChannelMode::Subscribe => flags::SUBSCRIBE,
-            ChannelMode::PresenceSubscribe => flags::PRESENCE_SUBSCRIBE,
-            ChannelMode::AnnotationPublish => flags::ANNOTATION_PUBLISH,
-            ChannelMode::AnnotationSubscribe => flags::ANNOTATION_SUBSCRIBE,
-        })
-        .fold(0, |acc, f| acc | f);
-    if ch.has_been_attached {
-        flag_bits |= flags::ATTACH_RESUME;
-    }
-    if flag_bits != 0 {
-        msg.flags = Some(flag_bits);
-    }
-    msg
-}
-
-/// RTL4m: decode the mode flags granted in ATTACHED.
-fn modes_from_flags(f: u64) -> Vec<ChannelMode> {
-    let mut modes = Vec::new();
-    if f & flags::PRESENCE != 0 {
-        modes.push(ChannelMode::Presence);
-    }
-    if f & flags::PUBLISH != 0 {
-        modes.push(ChannelMode::Publish);
-    }
-    if f & flags::SUBSCRIBE != 0 {
-        modes.push(ChannelMode::Subscribe);
-    }
-    if f & flags::PRESENCE_SUBSCRIBE != 0 {
-        modes.push(ChannelMode::PresenceSubscribe);
-    }
-    if f & flags::ANNOTATION_PUBLISH != 0 {
-        modes.push(ChannelMode::AnnotationPublish);
-    }
-    if f & flags::ANNOTATION_SUBSCRIBE != 0 {
-        modes.push(ChannelMode::AnnotationSubscribe);
-    }
-    modes
 }
 
 fn state_event(state: ConnectionState) -> ConnectionEvent {
